Remove iframelike priority for elbow arrows and refactor
