+++ packages/excalidraw/scene/comparisons.ts
@@ -1,8 +1,3 @@
-import { isIframeLikeElement } from "../element/typeChecks";
-import type {
-  ExcalidrawIframeLikeElement,
-  NonDeletedExcalidrawElement,
-} from "../element/types";
 import type { ElementOrToolType } from "../types";
 
 export const hasBackground = (type: ElementOrToolType) =>
@@ -47,52 +42,3 @@ export const canChangeRoundness = (type: ElementOrToolType) =>
 export const toolIsArrow = (type: ElementOrToolType) => type === "arrow";
 
 export const canHaveArrowheads = (type: ElementOrToolType) => type === "arrow";
-
-export const getElementAtPosition = (
-  elements: readonly NonDeletedExcalidrawElement[],
-  isAtPositionFn: (element: NonDeletedExcalidrawElement) => boolean,
-) => {
-  let hitElement = null;
-  // We need to to hit testing from front (end of the array) to back (beginning of the array)
-  // because array is ordered from lower z-index to highest and we want element z-index
-  // with higher z-index
-  for (let index = elements.length - 1; index >= 0; --index) {
-    const element = elements[index];
-    if (element.isDeleted) {
-      continue;
-    }
-    if (isAtPositionFn(element)) {
-      hitElement = element;
-      break;
-    }
-  }
-
-  return hitElement;
-};
-
-export const getElementsAtPosition = (
-  elements: readonly NonDeletedExcalidrawElement[],
-  isAtPositionFn: (element: NonDeletedExcalidrawElement) => boolean,
-) => {
-  const iframeLikes: ExcalidrawIframeLikeElement[] = [];
-  const elementsAtPosition: NonDeletedExcalidrawElement[] = [];
-  // We need to to hit testing from front (end of the array) to back (beginning of the array)
-  // because array is ordered from lower z-index to highest and we want element z-index
-  // with higher z-index
-  for (let index = elements.length - 1; index >= 0; --index) {
-    const element = elements[index];
-    if (element.isDeleted) {
-      continue;
-    }
-
-    if (isAtPositionFn(element)) {
-      if (isIframeLikeElement(element)) {
-        iframeLikes.push(element);
-      } else {
-        elementsAtPosition.push(element);
-      }
-    }
-  }
-
-  return elementsAtPosition.concat(iframeLikes);
-};
